data-plane/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis[hiredis]==5.0.1
celery==5.3.6
finnhub-python==2.4.19
httpx==0.26.0
structlog==24.1.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-prometheus==0.43b0
prometheus-client==0.19.0

